fixed use_mex=2 bug in pelvis control block

+++ examples/Atlas/controllers/FootContactBlock.m
@@ -149,7 +149,7 @@ classdef FootContactBlock < MIMODrakeSystem
           if isfield(ctrl_data.link_constraints(1),'pelvis_reference_height')
             pelvis_ref = ctrl_data.link_constraints(1).pelvis_reference_height(supp_idx);
             pelvis_ref_next = ctrl_data.link_constraints(1).pelvis_reference_height(supp_idx+1);
-            eta = (t1-t)/(t1-t0);
+            eta = double((t1-t)/(t1-t0));
             obj.controller_data.pelvis_foot_height_reference = eta * pelvis_ref + (1-eta)*pelvis_ref_next - ctrl_data.plan_shift(3); 
           end
         end
